✨ Add source examples for progress bars with Rich
