+++ docs_src/progressbar/tutorial001.py
@@ -1,15 +1,15 @@
 import time
 
 import typer
+from rich.progress import track
 
 
 def main():
     total = 0
-    with typer.progressbar(range(100)) as progress:
-        for value in progress:
-            # Fake processing time
-            time.sleep(0.01)
-            total += 1
+    for value in track(range(100), description="Processing..."):
+        # Fake processing time
+        time.sleep(0.01)
+        total += 1
     print(f"Processed {total} things.")
