+++ docs_src/progressbar/tutorial002.py
@@ -1,22 +1,19 @@
 import time
 
 import typer
-
-
-def iterate_user_ids():
-    # Let's imagine this is a web API, not a range()
-    for i in range(100):
-        yield i
+from rich.progress import Progress, SpinnerColumn, TextColumn
 
 
 def main():
-    total = 0
-    with typer.progressbar(iterate_user_ids(), length=100) as progress:
-        for value in progress:
-            # Fake processing time
-            time.sleep(0.01)
-            total += 1
-    print(f"Processed {total} user IDs.")
+    with Progress(
+        SpinnerColumn(),
+        TextColumn("[progress.description]{task.description}"),
+        transient=True,
+    ) as progress:
+        progress.add_task(description="Processing...", total=None)
+        progress.add_task(description="Preparing...", total=None)
+        time.sleep(5)
+    print("Done!")
 
 
 if __name__ == "__main__":
