+++ docs_src/progressbar/tutorial004.py
@@ -3,14 +3,20 @@ import time
 import typer
 
 
+def iterate_user_ids():
+    # Let's imagine this is a web API, not a range()
+    for i in range(100):
+        yield i
+
+
 def main():
-    total = 1000
-    with typer.progressbar(length=total) as progress:
-        for batch in range(4):
+    total = 0
+    with typer.progressbar(iterate_user_ids(), length=100) as progress:
+        for value in progress:
             # Fake processing time
-            time.sleep(1)
-            progress.update(250)
-    print(f"Processed {total} things in batches.")
+            time.sleep(0.01)
+            total += 1
+    print(f"Processed {total} user IDs.")
 
 
 if __name__ == "__main__":
