+++ docs_src/progressbar/tutorial005.py
@@ -5,7 +5,7 @@ import typer
 
 def main():
     total = 0
-    with typer.progressbar(range(100)) as progress:
+    with typer.progressbar(range(100), label="Processing") as progress:
         for value in progress:
             # Fake processing time
             time.sleep(0.01)
